nit style

+++ dracula.c
@@ -4,6 +4,7 @@
 //
 // Adam Yi <i@adamyi.com>, Simon Hanly-Jones <simon.hanly.jones@gmail.com>
 
+#include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -32,7 +33,6 @@
 #define DRAC_LOG AC_LOG_INFO
 #endif
 
-
 static double action_q[NUM_MAP_LOCATIONS], next_q[NUM_MAP_LOCATIONS];
 
 static inline int weighted_spdist(int spdist) {
@@ -59,9 +59,9 @@ static inline double weighted_cddist(int spdist, int health, int hDistToCP) {
   if (hDistToCP <= -200)
     weight = 0;
   else if (hDistToCP < 0)
-    weight = -1.2; // weight *= 0.01;
+    weight = -1.2;  // weight *= 0.01;
   else if (hDistToCP <= 12)
-    weight = -0.8; // weight *= 0.6;
+    weight = -0.8;  // weight *= 0.6;
   else if (hDistToCP > 16)
     weight *= 1.2;
   else if (hDistToCP > 32)
@@ -79,27 +79,34 @@ static inline double apply_weight(double x, double weight) {
   return (x / weight);
 }
 
-static inline void addQ(location_t loc, double action, double next, const char *reason) {
-  ac_log(DRAC_LOG, "add (%lf,%lf) to Q[%s] for %s", action, next, location_get_abbrev(loc), reason);
+static inline void addQ(location_t loc, double action, double next,
+                        const char *reason) {
+  ac_log(DRAC_LOG, "add (%lf,%lf) to Q[%s] for %s", action, next,
+         location_get_abbrev(loc), reason);
   action_q[loc] += action;
   next_q[loc] += next;
 }
 
-static inline void setQ(location_t loc, double action, double next, const char *reason) {
-  ac_log(DRAC_LOG, "set (%lf,%lf) to Q[%s] for %s", action, next, location_get_abbrev(loc), reason);
+static inline void setQ(location_t loc, double action, double next,
+                        const char *reason) {
+  ac_log(DRAC_LOG, "set (%lf,%lf) to Q[%s] for %s", action, next,
+         location_get_abbrev(loc), reason);
   action_q[loc] = action;
   next_q[loc] = next;
 }
 
-static inline void applyWeightQ(location_t loc, double action, double next, const char *reason) {
-  ac_log(DRAC_LOG, "apply weight (%lf,%lf) to Q[%s] for %s", action, next, location_get_abbrev(loc), reason);
+static inline void applyWeightQ(location_t loc, double action, double next,
+                                const char *reason) {
+  ac_log(DRAC_LOG, "apply weight (%lf,%lf) to Q[%s] for %s", action, next,
+         location_get_abbrev(loc), reason);
   action_q[loc] = apply_weight(action_q[loc], action);
   next_q[loc] = apply_weight(next_q[loc], next);
 }
 
 static inline void printQMatrix(location_t *possible, size_t sz) {
   for (size_t i = 0; i < sz; i++)
-    ac_log(DRAC_LOG, "%s: %lf %lf", location_get_abbrev(possible[i]), action_q[possible[i]], next_q[possible[i]]);
+    ac_log(DRAC_LOG, "%s: %lf %lf", location_get_abbrev(possible[i]),
+           action_q[possible[i]], next_q[possible[i]]);
 }
 
 void decide_dracula_move(DraculaView dv) {
@@ -123,7 +130,8 @@ void decide_dracula_move(DraculaView dv) {
   for (size_t i = 0; i < num; i++) {
     if (possible[i] >= HIDE) {
       location_t res = player_resolve_move_location(dracp, possible[i]);
-      if (possible[i] == HIDE && cango[res] && rev[res] >= DOUBLE_BACK_1) // prefer DB to HIDE
+      if (possible[i] == HIDE && cango[res] &&
+          rev[res] >= DOUBLE_BACK_1)  // prefer DB to HIDE
         continue;
       rev[res] = possible[i];
       possible[i] = res;
@@ -151,48 +159,59 @@ void decide_dracula_move(DraculaView dv) {
       int dst = SPDIST[i][hl] + SPDIST[CASTLE_DRACULA][hl] - distToCD;
       if (dst < md) md = dst;
       int hunter_max_trail = (round + 1 + i) % 4;
-      addQ(i, weighted_spdist(SPROUND[hl][i][hunter_max_trail]), weighted_spdist(fmax(SPROUND[hl][i][hunter_max_trail] - 1, 0)), "distance to hunter");
+      addQ(i, weighted_spdist(SPROUND[hl][i][hunter_max_trail]),
+           weighted_spdist((int)fmax(SPROUND[hl][i][hunter_max_trail] - 1, 0)),
+           "distance to hunter");
     }
     if (md == 0) {
-      addQ(i, apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.5), apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.5), "how good to go to CD (*0.5)");
+      addQ(i, apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.5),
+           apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.5),
+           "how good to go to CD (*0.5)");
     } else if (md == 1) {
-      addQ(i, apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.7), apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.7), "how good to go to CD (*0.7)");
+      addQ(i, apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.7),
+           apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.7),
+           "how good to go to CD (*0.7)");
     } else if (md == 2) {
-      addQ(i, apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.8), apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.8), "how good to go to CD (*0.8)");
+      addQ(i, apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.8),
+           apply_weight(weighted_cddist(distToCD, health, hDistSp), 0.8),
+           "how good to go to CD (*0.8)");
     } else {
-      addQ(i, weighted_cddist(distToCD, health, hDistSp), weighted_cddist(distToCD, health, hDistSp), "how good to go to CD (*1.0)");
+      addQ(i, weighted_cddist(distToCD, health, hDistSp),
+           weighted_cddist(distToCD, health, hDistSp),
+           "how good to go to CD (*1.0)");
     }
     if (location_get_type(i) == SEA) {
       if (action_q[i] < 0) {
         if (health <= 8)
-          applyWeightQ(i, 1.0/0.8, 1.0/0.8, "will die soon on SEA (hunter close by)");
+          applyWeightQ(i, 1.0 / 0.8, 1.0 / 0.8,
+                       "will die soon on SEA (hunter close by)");
         else if (health > 20)
-          applyWeightQ(i, 1.0/0.1, 1.0/0.1, "plenty of life on SEA (hunter close by)");
+          applyWeightQ(i, 1.0 / 0.1, 1.0 / 0.1,
+                       "plenty of life on SEA (hunter close by)");
         else
-          applyWeightQ(i, 1.0/0.35, 1.0/0.35, "plenty of life on SEA (hunter close by)");
+          applyWeightQ(i, 1.0 / 0.35, 1.0 / 0.35,
+                       "plenty of life on SEA (hunter close by)");
         // action_q[i] -= 3;
       } else if (round % 13 == 0 || round % 13 == 12)
         applyWeightQ(i, 0.25, 0.25, "unable to place vampire on SEA");
       else
-        applyWeightQ(i, NORMAL_SEA_WEIGHT, NORMAL_SEA_WEIGHT, "we don't favor SEA (no hunter nearby)");
-      if (health <= 2)
-        setQ(i, -100, -100, "will die on SEA");
+        applyWeightQ(i, NORMAL_SEA_WEIGHT, NORMAL_SEA_WEIGHT,
+                     "we don't favor SEA (no hunter nearby)");
+      if (health <= 2) setQ(i, -100, -100, "will die on SEA");
     }
-    if (cango[i] && rev[i] != i) // DB or HIDE
+    if (cango[i] && rev[i] != i)  // DB or HIDE
       applyWeightQ(i, DB_HIDE_WEIGHT, 1, "we don't favor DB/HIDE");
   }
   printQMatrix(possible, num);
   for (int i = MIN_MAP_LOCATION; i <= MAX_MAP_LOCATION; i++) {
-    if (!cango[i])
-      continue;
+    if (!cango[i]) continue;
     player_t *tmpp = clone_player(dracp);
     tmpp->all_history_size = -1;
     player_move_to(tmpp, i, rev[i]);
     size_t tmps = 0;
-    location_t *tmpl = _gv_do_get_connections(
-        tmpp, &tmps,
-        i, PLAYER_DRACULA, round + 1,
-        true, false, true, true, false, true);
+    location_t *tmpl =
+        _gv_do_get_connections(tmpp, &tmps, i, PLAYER_DRACULA, round + 1, true,
+                               false, true, true, false, true);
     double bstNextMove = -10000;
     size_t nxtMoves = tmps;
     for (int j = 0; j < tmps; j++) {
@@ -203,26 +222,29 @@ void decide_dracula_move(DraculaView dv) {
         l = player_resolve_move_location(tmpp, l);
         w = DB_HIDE_WEIGHT;
         if (next_q[CASTLE_DRACULA] < 0)
-            w = -1; // we don't want to teleport (e.g. forced to hide then db then tp)
+          w = -1;  // we don't want to teleport (e.g. forced to hide then db
+                   // then tp)
       } else if (l == TELEPORT) {
         l = CASTLE_DRACULA;
         if (dracp->location == CASTLE_DRACULA)
-          w = -1; // we don't want to be locked in at CD
+          w = -1;  // we don't want to be locked in at CD
       }
       w = apply_weight(next_q[l], w);
       printf("%s %lf\n", location_get_abbrev(l), w);
-      if (w > bstNextMove)
-        bstNextMove = w;
+      if (w > bstNextMove) bstNextMove = w;
     }
     printf("%s bstNextMove %lf\n", location_get_abbrev(i), bstNextMove);
     free(tmpl);
     destroy_player(tmpp);
     addQ(i, NEXT_Q_GAMMA * (bstNextMove - action_q[i]), 0, "next step bias");
-    // if (nxtMoves >= 4) addQ(i, 0.1 * (nxtMoves - 4) + 1, 0, "next step action space reward");
+    // if (nxtMoves >= 4) addQ(i, 0.1 * (nxtMoves - 4) + 1, 0, "next step action
+    // space reward");
     if (action_q[i] < 0)
-      addQ(i, 0.5 * nxtMoves, 0, "next step action space reward (hunter closeby)");
+      addQ(i, 0.5 * nxtMoves, 0,
+           "next step action space reward (hunter closeby)");
     else if (nxtMoves >= 4)
-      addQ(i, 0.1 * (nxtMoves - 4) + 1, 0, "next step action space reward (hunter far away)");
+      addQ(i, 0.1 * (nxtMoves - 4) + 1, 0,
+           "next step action space reward (hunter far away)");
   }
 
   double maxdist = -10000;
